# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon API
  version: "1.0.0"
  description: |-
    Beacon is a crisis-damage crowdsourcing backend: anonymous mobile reporters
    submit building-damage reports (identified only by an `X-Device-Id` header),
    and authenticated analysts (JWT bearer) verify, dispatch and export them.

    ## Visibility tiers
    - **Anonymous / public** callers (no JWT) and the authenticated read-only
      `external_viewer` role see **verified reports only**, in a locked-down
      *public projection*: coordinates coarsened to 3 decimal places (~110 m),
      and `submitterId`, `what3words`, `plusCode`, `landmark`, `buildingId`,
      `buildingSource`, `infraName`, `gpsAccuracyMeters`, `description`,
      tasking/operational fields, `modular` and `anonymization` are stripped.
      Unverified reports return 404 so their existence and position are never
      leaked. Public aggregates (`/reports/area-groups`) count verified
      reports only for this tier, coherent with the verified-only public map.
    - **Analyst roles** (`field_validator`, `co_analyst`, `regional_analyst`,
      `crisis_admin`) see all statuses at full precision, limited to the
      crises in their `crisisScope` (`"*"` = org-wide).

    ## Crisis scoping
    Where `crisisId` is optional (stats, export), an omitted value resolves to
    the **newest active crisis** (the same crisis `GET /v1/crises/active`
    returns), falling back to a seeded default only when no crisis is active.

...

...

...
servers:
...
paths:
...
  /v1/crises/{id}/form:
    patch:
      tags: [forms]
      operationId: patchCrisisForm
      summary: Adjust a crisis's modular capture form (senior analysts)
      description: |-
        Stores per-crisis overrides over the default Appendix-1 sections:
        section keys in `required` are flipped to required, keys in `disabled`
        are removed from the form entirely. Keys must name known sections
        (`electricity`, `healthServices`, `pressingNeeds`) and a key cannot be
        in both lists. Sending two empty lists resets the crisis to the pure
        defaults.

        **RBAC:** `regional_analyst` / `crisis_admin` only (shaping what every
        reporter in a crisis is asked is a senior decision, like
        `/crises/{id}/status`), and the crisis must be inside the caller's
        `crisisScope`. Other mutator roles get 403 `forbidden`; out-of-scope
        gets 403 `out_of_scope`.

        Responds with the crisis's RESOLVED schema (same shape as
        `GET /v1/form-schema`) so a dashboard can re-render immediately.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormOverrides'
      responses:
        "200":
          description: The crisis's resolved form schema after the update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormSchema'
        "400":
          description: Malformed JSON (`bad_json`), unknown section key or a key in both lists (`validation`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/OutOfScope'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/form-schema:
    get:
      tags: [forms]
      operationId: getFormSchema
      summary: The modular capture-form schema for a crisis (public)
      description: |-
        PUBLIC (like `/v1/crises`): the anonymous mobile capture flow downloads
        the modular form definition before submitting. Returns the built-in
        Appendix-1 sections — electricity (6 options), healthServices (5),
        pressingNeeds (10, multi-select with an "Other → specify" free text) —
        with the crisis's stored overrides applied: `required` flips a section
        to required, `disabled` removes it.

        Option `value`s are EXACTLY the wire values the mobile client already
        stores in the report's `modular` blob (the `ModularImpacts` enums);
        sections default to optional (the modular framing — the required core
        indicators are fixed report fields, not these).

        With no `crisisId` the newest active crisis's form is returned (same
        default scope as the other public reads); an unknown crisis id
        resolves to the plain defaults rather than an error, so offline-first
        clients holding a stale id always get a usable form.
      parameters:
        - name: crisisId
          in: query
          schema: { type: string }
          description: The crisis whose form to resolve (default = newest active crisis).
      responses:
        "200":
          description: The resolved form sections.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormSchema'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Analyst JWT issued by POST /v1/auth/login.
...
  responses:
...
    Unauthorized:
      description: Analyst login required (missing/invalid bearer token).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    OutOfScope:
      description: The caller's role/scope cannot access or modify this resource (`out_of_scope` / `forbidden`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found (also returned for resources the caller's tier may not see).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Stable machine-readable code (e.g. validation, bad_json, unauthorized, out_of_scope, not_found, rate_limited).
        message:
          type: string
...
    FormOption:
      type: object
      required: [value, label]
      properties:
        value:
          type: string
          description: The wire value stored in the report's `modular` blob (matches the `ModularImpacts` enums — never re-invented).
        label:
          type: string
          description: The Appendix-1 display text.
    FormSection:
      type: object
      required: [key, title, type, required, options]
      properties:
        key:
          type: string
          enum: [electricity, healthServices, pressingNeeds]
          description: The section's key in the report's `modular` blob.
        title: { type: string }
        type:
          type: string
          enum: [single, multi]
        required:
          type: boolean
          description: Defaults to false (modular framing); flipped per crisis via form overrides.
        allowOtherText:
          type: boolean
          description: Present (true) on multi sections with an "Other → specify" free text (stored as `pressingNeedsOther`).
        options:
          type: array
          items:
            $ref: '#/components/schemas/FormOption'
    FormSchema:
      type: object
      required: [sections]
      properties:
        sections:
          type: array
          items:
            $ref: '#/components/schemas/FormSection'
    FormOverrides:
      type: object
      description: |-
        Per-crisis form adjustment (the PATCH /v1/crises/{id}/form body and the
        stored `crises.form_overrides` document). Keys must name known
        sections; a key cannot appear in both lists. Omitted lists are treated
        as empty; two empty lists reset the crisis to the defaults.
      properties:
        required:
          type: array
          items: { type: string }
          description: Section keys flipped to required.
        disabled:
          type: array
          items: { type: string }
          description: Section keys removed from the form entirely.
...